refactor(cypress): split lib.rs into lib, scanner, helpers, and tests modules
Same shape as the storybook refactor: the 1816-line crates/cypress/src/lib.rs
is split into four modules. The Scanner struct + impl moves to scanner.rs,
free helpers move to helpers.rs, tests move to tests.rs, and lib.rs keeps
only the public API surface. All three existing unit tests pass.
